backend/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.4
pycoingecko==2.2.0
PyJWT==2.3.0
python-dotenv==0.15.0
pytz==2021.3
requests==2.26.0
six==1.15.0
twilio==7.1.0
urllib3==1.26.7
Werkzeug==1.0.1
